Code commenté + doc

+++ LapinManager.cpp
@@ -1,3 +1,16 @@
+/**
+ *
+ *	\file 		LapinManager.cpp
+ *	\brief 		Simulateur de population de lapins.
+ *	\author 	Benoit GARCON
+ *  \author     Benjamin BARBESANGE
+ *	\version 	1.0
+ *	\date 		20 Novembre 2015
+ *
+ *	Simulateur de population de lapins à partir d'un couple de un an
+ *
+ */
+
 #include "LapinManager.h"
 
 using namespace std;
@@ -6,9 +19,11 @@ LapinManager::LapinManager(std::string p_str) :
     m_graine(p_str.begin(), p_str.end()), m_generateur(m_graine),
     m_mois(0), m_males(180,ClasseLapins(&m_generateur)), m_femelles(180,ClasseLapins(&m_generateur)) {  // 15 * 12
 
+	// ajout d'Adam et Eve ou bien Bob et Alice
 	m_males[12].nombre(1);
 	m_femelles[12].nombre(1);
 
+	// initialisation des taux de survie "speciaux"
 	for(int i = 0 ; i < 9 ; ++i) {
 		m_males[i].tauxDeSurvie(pow(0.2,((double)1/12)));
 		m_femelles[i].tauxDeSurvie(pow(0.2,((double)1/12)));
@@ -17,24 +32,26 @@ LapinManager::LapinManager(std::string p_str) :
 		m_males[i].tauxDeSurvie(pow(0.4-0.1*((int)(i-132)/12), ((double)1/12)));
 		m_femelles[i].tauxDeSurvie(pow(0.4-0.1*((int)(i-132)/12),((double)1/12)));
 	}
+	// ouverture du fichier de sortie
 	m_file.open ("lap.out", std::ofstream::out | std::ofstream::trunc);
 }
 
 LapinManager::~LapinManager() {
+	// fermeture du fichier de sortie
 	m_file.close();
 }
 
-long double LapinManager::simulation(unsigned long long p_duree, bool p_print, bool p_outFile) { // p_print indique si on a affiche ou pas
-    long double             population = 0,
-							dateStop = m_mois+p_duree;
-	bool					writeInFile = m_file.is_open() && p_outFile;
-    for(/*m_mois*/ ; m_mois < dateStop ; ++m_mois) {
-        /// mort veillissement
-        long double             nouvelEffectifMale      = 0,
-                                nouvelEffectifFemelle   = 0,
-                                males   				= 0,
-                                femelles  	 			= 0,
-								morts 					= 0;
+INTEGER LapinManager::simulation(unsigned long long p_duree, bool p_print, bool p_outFile) { // p_print indique si on a affiche ou pas
+    INTEGER					population = 0,									// Compteur de population totale
+							dateStop = m_mois+p_duree;						// Date d'arret
+	bool					writeInFile = m_file.is_open() && p_outFile;	// Booleen vrai alors ecriture dans un fichier
+    for(/*m_mois*/ ; m_mois < dateStop ; ++m_mois) {	// pour chaque mois jusqu'a la date d'arret
+        // mort veillissement
+        INTEGER					nouvelEffectifMale      = 0,		// Nombre de males d'une classe survivant
+                                nouvelEffectifFemelle   = 0,		// Nombre de femelles d'une classe survivant
+                                males   				= 0,		//	Nombre de males total
+                                femelles  	 			= 0,		// Nombre de femelles	total
+								morts 					= 0;		// Nombre de morts mensuel
         population = 0;
         for(unsigned i = 0 ; i < m_males.size() ; ++i) {
 			males += nouvelEffectifMale;
@@ -46,21 +63,21 @@ long double LapinManager::simulation(unsigned long long p_duree, bool p_print, b
 			morts = morts - (nouvelEffectifMale+nouvelEffectifFemelle);
         }
 
-        /// naissances
-        long double             naissances = 0,
-                                bebeMales = 0;
+        // naissances
+        INTEGER					naissances = 0,		// Nombre de naissances mensuel
+                                bebeMales = 0;		// Nombre de nouveaux nes males
 
         for(unsigned i = 9 ; i < m_femelles.size() ; ++i) {         // nombre de naissances le mois
             naissances += m_femelles[i].reproduction(m_mois%12);
         }
-        /// determination du sexe
+        // determination du sexe
         if(naissances >= 100) {           // traitement pour grand nombre de naissance
-            std::normal_distribution<long double>   normale(0.5*naissances,
-                                                    sqrt(0.25*naissances));
+            std::normal_distribution<long double>	normale(0.5*naissances,
+                                                    sqrt(0.25*naissances));		// Loi normale
             bebeMales = round(normale(m_generateur));
         } else {                        // traitement individuel
             for(unsigned i = 0 ; i < naissances ; ++i) {
-                std::bernoulli_distribution bernoulli(0.5);
+                std::bernoulli_distribution bernoulli(0.5);		// Loi Bernouilli
                 bebeMales += (bernoulli(m_generateur) ? 1 : 0);
             }
         }
@@ -70,7 +87,7 @@ long double LapinManager::simulation(unsigned long long p_duree, bool p_print, b
 		femelles += m_femelles[0].nombre();
         population += naissances;
 
-        if(p_print) {
+        if(p_print) {	// affichage console
         	cout << "[ " << (m_mois%12)+1 << "/" << 2000+(m_mois/12) << " ]\t--------------------------------" << endl;
 			cout << "Nombre d'individus :\t" << population << endl;
         	cout << "Nombre de naissances :\t" << naissances << endl;
@@ -78,7 +95,7 @@ long double LapinManager::simulation(unsigned long long p_duree, bool p_print, b
         	cout << "Nombre de males :\t" << males << endl;
         	cout << "Nombre de femelles :\t" << femelles << endl;
         }
-		if(writeInFile) {
+		if(writeInFile) {	// sortie fichier
 			m_file << "[ " << (m_mois%12)+1 << "/" << 2000+(m_mois/12) << " ]\t" << population << "\t" << naissances << "\t" << morts << "\t" << males << "\t" << femelles << endl;
 		}
     }
@@ -87,13 +104,13 @@ long double LapinManager::simulation(unsigned long long p_duree, bool p_print, b
 
 void LapinManager::reset() { // remise a zero pour refaire une simu
     m_mois = 0;
-    /// RAZ du nombre de lapins
+    // RAZ du nombre de lapins
     for(unsigned i = 0; i < m_males.size(); i++) {
         m_males[i].nombre(0);
         m_femelles[i].nombre(0);
     }
 
-    /// Ajout de 2 individus
+    // Ajout de 2 individus
     m_males[12].nombre(1);
     m_femelles[12].nombre(1);
 }
